Development enhancements.

+++ scripts/run_backtest.sh
@@ -3,15 +3,16 @@ CWD="$(cd -P -- "$(dirname -- "$0")" && pwd -P)"
 ARGS=":r:e:f:E:c:p:d:m:y:s:oi:I:Cb:tTD:vxh"
 
 ## Check dependencies.
-type git ex
+type git ex xdpyinfo
 
 ## Define functions.
+
 # Invoke on test success.
 on_success() {
   ! grep -C2 -e "Initialization failed" <(show_logs) || exit 1
   echo "Printing logs..." >&2
   show_logs
-  echo "TEST succeded." >&2
+  echo "TEST succeeded." >&2
   parse_results $@
   on_finish
   local OPTIND
@@ -28,6 +29,10 @@ on_success() {
 
 # Invoke on test failure.
 on_failure() {
+  # Sometimes MT4 fails on success, therefore double checking.
+  REPORT_HTM=$(find "$TESTER_DIR" -name "$(basename "$(ini_get TestReport)").htm")
+  test -f "$REPORT_HTM" && on_success $@
+
   echo "Printing logs..." >&2
   show_logs
   echo "TEST failed." >&2
@@ -47,19 +52,19 @@ parse_results() {
   REPORT_BASE="$(basename "$(ini_get TestReport)")"
   REPORT_HTM=$(find "$TESTER_DIR" -name "${REPORT_BASE}.htm")
   test -f "$REPORT_HTM" || exit 1
-  echo "Checking and saving time..." >&2
+  echo "Checking the total time elapsed..." >&2
   save_time
   while getopts $ARGS arg; do
     case $arg in
       t) # Convert test report file into brief text format.
         echo "Converting report into short text file..."
         REPORT_TXT="$(dirname "$REPORT_HTM")/$REPORT_BASE.txt"
-        html2text -width 105 "$REPORT_HTM" | sed "/\[Graph\]/q" > "$REPORT_TXT" && rm -v "$REPORT_HTM"
+        grep -v mso-number "$REPORT_HTM" | html2text -width 105 | sed "/\[Graph\]/q" > "$REPORT_TXT" && rm -v "$REPORT_HTM"
         ;;
       T) # Convert test report file into full detailed text format.
         echo "Converting report into text file..."
         REPORT_TXT="$(dirname "$REPORT_HTM")/$REPORT_BASE.txt"
-        html2text -width 105 -o "$REPORT_TXT" "$REPORT_HTM" && rm -v "$REPORT_HTM"
+        grep -v mso-number "$REPORT_HTM" | html2text -width 105 -o "$REPORT_TXT" && rm -v "$REPORT_HTM"
         ;;
       D)
         echo "Copying report files..."
@@ -69,7 +74,7 @@ parse_results() {
         ;;
       v)
         echo "Printing test reports..."
-        html2text -width 180 "$REPORT_HTM" | sed "/\[Graph\]/q"
+        grep -v mso-number "$REPORT_HTM" | html2text -width 180 | sed "/\[Graph\]/q"
         find "$TESTER_DIR/files" '(' -name "*.log" -o -name "*.txt" ')' $VPRINT -exec cat "{}" +
         ;;
       o)
@@ -120,12 +125,30 @@ while getopts $ARGS arg; do
     e) # EA name.
       EA_NAME=${OPTARG}
       EA_PATH="$(find "$ROOT" '(' -name "*$EA_NAME*.ex4" -o -name "*$EA_NAME*.ex5" ')' -print -quit)"
-      [ -f "$EA_PATH" ]
+      [ -f "$EA_PATH" ] || { echo "Error: EA file ($EA_NAME) not found in '$ROOT'!" >&2; exit 1; }
       cp $VFLAG "$EA_PATH" "$TERMINAL_DIR/MQL4/Experts";
       ini_set "^TestExpert" "$(basename "${EA_PATH%.*}")" "$TESTER_INI"
       ;;
-    m) # How many months to test.
-      MONTHS=${OPTARG}
+    m) # Which months to test (default: 1-12)
+      IFS='-' MONTHS=(${OPTARG})
+      IFS=$' \t\n' # Restore IFS.
+      ;;
+
+    C) # Clear previous backtest data files.
+      clean_files
+      clean_bt
+      ;;
+
+    b) # Backtest data to test.
+      BT_SRC=${OPTARG}
+      echo "Checking backtest data ($BT_SRC)..."
+      bt_key="${SYMBOL:-EURUSD}-${YEAR:-2014}-${BT_SRC:-N1}"
+      # Generate backtest files if not present.
+      if [ ! -s "$(find "$TERMINAL_DIR" -name '*.fxt' -print -quit)" ] || [ "$(ini_get "bt_data" "$CUSTOM_INI")" != "$bt_key" ]; then
+        $SCR/get_bt_data.sh ${SYMBOL:-EURUSD} ${YEAR:-2014} ${BT_SRC:-N1}
+      else
+        echo "Skipping, as $bt_key already exists..."
+      fi
       ;;
   esac
 done
@@ -135,93 +158,85 @@ EA_NAME="$(ini_get TestExpert)"
 EA_INI="$TESTER_DIR/$EA_NAME.ini"
 cp $VFLAG "$TPL_EA" "$EA_INI"
 
+# Assign variables.
+FXT_FILE=$(find "$TICKDATA_DIR" -name "*.fxt" -print -quit)
+
 # Parse the secondary arguments.
 OPTIND=1
 while getopts $ARGS arg; do
   case ${arg} in
     r) # The name of the test report file.
-      echo "Setting test report..."
       REPORT="tester/$(basename "${OPTARG}")"
+      echo "Setting test report ($REPORT)..." >&2
       ini_set "^TestReport" "$REPORT" "$TESTER_INI"
       ;;
 
     f) # The .set file to run the test.
-      echo "Setting EA parameters..."
       SETORG="$OPTARG"
       SETFILE="${EA_NAME}.set"
+      echo "Setting EA parameters ($SETFILE)..." >&2
       [ -f "$SETORG" ]
       cp -f $VFLAG "$OPTARG" "$TESTER_DIR/$SETFILE"
       ini_set "^TestExpertParameters" "$SETFILE" "$TESTER_INI"
+      ini_set_inputs "$TESTER_DIR/$SETFILE" "$EA_INI"
       ;;
 
     E) # EA settings (e.g. genetic=0, maxdrawdown=20.00).
-      echo "Applying EA settings..."
       EA_OPTS=${OPTARG}
-      IFS='=' ea_option=($in)
+      echo "Applying EA settings ($EA_OPTS)..." >&2
+      IFS='='; ea_option=($EA_OPTS)
+      IFS=$' \t\n' # Restore IFS.
       [ -f "$EA_INI" ]
       ini_set "^${option[0]}" "${option[1]}" "$EA_INI"
       ;;
 
     c) # Base currency for test (e.g. USD).
       CURRENCY=${OPTARG}
+      echo "Setting base currency to $CURRENCY..." >&2
       ini_set "^currency" "$CURRENCY" "$EA_INI"
       ;;
 
     p) # Symbol pair to test (e.g. EURUSD).
-      echo "Setting symbol pair..."
       SYMBOL=${OPTARG}
+      echo "Setting symbol pair to $SYMBOL..." >&2
       ini_set "^TestSymbol" "$SYMBOL" "$TESTER_INI"
       ;;
 
     d) # Deposit amount to test (e.g. 2000).
-      echo "Setting deposit..."
       DEPOSIT=${OPTARG}
+      echo "Setting deposit to $DEPOSIT..." >&2
       ini_set "^deposit" "$DEPOSIT" "$EA_INI"
       ;;
 
     y) # Year to test (e.g. 2014).
-      echo "Setting period to test..."
       YEAR=${OPTARG}
-      ini_set "^TestFromDate" "$YEAR.01.01" "$TESTER_INI"
-      ini_set "^TestToDate"   "$YEAR.${MONTHS:-12}.30" "$TESTER_INI"
+      START_DATE="$YEAR.${MONTHS[0]:-01}.01"
+      END_DATE="$YEAR.${MONTHS[1]:-$(echo ${MONTHS[0]:-12})}.30"
+      echo "Setting test period ($START_DATE-$END_DATE)..." >&2
+      ini_set "^TestFromDate" "$START_DATE" "$TESTER_INI"
+      ini_set "^TestToDate"   "$END_DATE" "$TESTER_INI"
       ;;
 
     s) # Spread to test.
-      echo "Setting spread to test..."
       SPREAD=${OPTARG}
-      ini_set "^Spread" "$SPREAD" "$TERMINAL_INI"
-      ini_set "^TestSpread" "$SPREAD" "$TESTER_INI"
+      echo "Setting spread to test ($SPREAD)..." >&2
+      set_spread $SPREAD
       ;;
 
     o) # Run optimization test.
-      echo "Setting optimization mode..."
       OPTIMIZATION=true
+      echo "Setting optimization mode..." >&2
       ini_set "^TestOptimization" true "$TESTER_INI"
       ;;
 
     i) # Invoke file with custom rules.
-      echo "Invoking includes..."
+      type bc
       INCLUDE=${OPTARG}
       SETFILE="$(ini_get TestExpert).set"
-      [ -f "$TESTER_DIR/$SETFILE" ] || { echo "Please specify .set file first (-f)."; exit 1; }
+      [ -f "$TESTER_DIR/$SETFILE" ] || { echo "Please specify .set file first (-f)." >&2; exit 1; }
+      echo "Invoking include file ($INCLUDE)..."
       . "$INCLUDE"
-      ;;
-
-    C) # Clear previous backtest data files.
-      clean_files
-      clean_bt
-      ;;
-
-    b) # Backtest data to test.
-      echo "Checking backtest data..."
-      BT_SRC=${OPTARG}
-      bt_key="${SYMBOL:-EURUSD}-${YEAR:-2014}-${BT_SRC:-N1}"
-      # Generate backtest files if not present.
-      if [ ! -s "$(find "$TERMINAL_DIR" -name '*.fxt' -print -quit)" ] || [ "$(ini_get "bt_data" "$CUSTOM_INI")" != "$bt_key" ]; then
-        $SCR/get_bt_data.sh ${SYMBOL:-EURUSD} ${YEAR:-2014} ${BT_SRC:-N1}
-      else
-        echo "Skipping, as $bt_key already exists..."
-      fi
+      ini_set_inputs "$TESTER_DIR/$SETFILE" "$EA_INI"
       ;;
 
     t)
@@ -230,16 +245,16 @@ while getopts $ARGS arg; do
 
     D) # Destination directory to save the test results.
       DEST=${OPTARG}
-      echo "Checking destination ..."
+      echo "Checking destination ($DEST)..." >&2
       [ -d "$DEST" ] || mkdir -p "$DEST"
       ;;
 
     # Placeholders for parameters used somewhere else.
-    m | e | I | v | x) ;;
+    e | m | C | b | I | v | x) ;;
 
     \? | h | *) # Display help.
-      echo "$0 usage:"
-      grep " .)\ #" $0
+      echo "$0 usage:" >&2
+      grep " .)\ #" $0 >&2
       exit 0
       ;;
 
@@ -252,4 +267,5 @@ done
 clean_files
 
 # Run the test under the platform.
-(time wine "$TERMINAL_EXE" "config/$CONF_TEST") 2> "$TERMINAL_LOG" && on_success $@ || on_failure
+configure_display
+(time wine "$TERMINAL_EXE" "config/$CONF_TEST") 2> "$TERMINAL_LOG" && on_success $@ || on_failure $@
